Move webmentions/location/extract code to indieweb application
